+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend/kubernetes_kurtosis_backend_engine_functions.go
@@ -60,6 +60,7 @@ func (backend KubernetesKurtosisBackend) CreateEngine(
 	grpcPortNum uint16,
 	grpcProxyPortNum uint16,
 	envVars map[string]string,
+	logsDbVolumeDirpath string,
 ) (
 	*engine.Engine,
 	error,