+++ container-engine-lib/docs/changelog.md
@@ -4,6 +4,8 @@
 * Removed duplicated shared functions that where present in the both the `docker_kurtosis_backend` package and the `shared_helpers` package
 * Added the `consts` package to group the general constants used in the Docker Kurtosis backend
 * Added the `ForLogsDatabaseServer` and the `ForLogsCollectorServer` Docker objects attribute
+* Added `LogsCollectorTypeDockerLabelValue`, `ModuleContainerTypeDockerLabelValue` and `LogsDbDataVolumeTypeDockerLabelValue` Docker labels
+* Added `EngineGUIDDockerLabelKey` Docker label key
 
 #### Breaking Changes
 * Move all `docker.DockerKurtosisBackend` functionality to `docker_kurtosis_backend` module